Firebase Unit Suite testing for Database, and Set up unit test for sign up
I added the firebase emulator implementation recently in a new directory. The use for this is to test the database locally and check if our function calls we make to the firebase api work properly. Let me know if you guys need a tutorial for setting this up.

Another thing I added was finally being able to set up the Unit testing for signup. The undefined issue we were getting was fixed by using the mock functionality. Just be sure to pass the firebase api key when you guys need to unit test. Still haven't found a fix for that issue yet. Let me know what you guys think could be improved or what else I could possibly test for SignUp .